agent: fix clippy findings in the integrated copyfile code
Two unnecessary u32 casts in the new agent tests, and ShareFsVolume::new now
takes eight arguments after gaining copy_volume_types. The argument count is
inherent to the constructor rather than a sign it should be split, so it is
allowed explicitly.

+++ src/agent/src/rpc.rs
@@ -4654,7 +4654,7 @@ mod tests {
             agent_volume_id: agent_volume_id.to_string(),
             file_name: file_name.to_string(),
             file_size: data.len() as i64,
-            file_mode: 0o644 | libc::S_IFREG as u32,
+            file_mode: 0o644 | libc::S_IFREG,
             uid: unistd::getuid().as_raw() as i32,
             gid: unistd::getgid().as_raw() as i32,
             offset: 0,
@@ -4711,7 +4711,7 @@ mod tests {
                     gid: unistd::getgid().as_raw() as i32,
                     data_size: data.len() as i64,
                     data,
-                    file_mode: 0o644 | libc::S_IFREG as u32,
+                    file_mode: 0o644 | libc::S_IFREG,
                     ..Default::default()
                 },
             )
